+++ src/routes/position.js
@@ -4,6 +4,7 @@ const positionEligibilityModel = require("../models/positionModel");
 const studentAuth = require("../middleware/studentAuth");
 const companyData = require("../models/newCompanyModel");
 const logger = require("./logger");
+const jwt = require("jsonwebtoken");
 const {
   client,
   postionCacheData,
@@ -42,24 +43,24 @@ const {
  *                             type : string
  *                      additionalCriteria :
  *                             type : string
- *                      degrees : 
+ *                      degrees :
  *                             type :  [string]
- *                      streams : 
+ *                      streams :
  *                             type :  [string]
  *                      graduationsYear :
  *                             type :  number
- *                      locationDomiciles : 
+ *                      locationDomiciles :
  *                             type :  [string]
- *                      tenthPer : 
- *                             type :  number 
- *                      gender : 
- *                             type :  string 
- *                             enum : 
+ *                      tenthPer :
+ *                             type :  number
+ *                      gender :
+ *                             type :  string
+ *                             enum :
  *                              - Male
- *                              - Female 
+ *                              - Female
  *                              - Other
- * 
- *                      
+ *
+ *
  */
 
 /**
@@ -77,23 +78,39 @@ const {
  *            description: Internet server problem
  *
  */
-positionRoute.get("/position", postionCacheData,async (req, res) => {
+positionRoute.get("/position", postionCacheData, async (req, res) => {
   try {
-    const positionEligibilityData = await positionEligibilityModel.find({});
-    if (positionEligibilityData.length <= 0) {
-     return res.status(404).send({ message: "data not available" });
+    const token = req.headers["authorization"]?.split(" ")[1];
+
+    if (!token) {
+      logger.error("No token provided");
+      return res.status(401).send({ message: "Unauthorized" });
     }
-    client.setEx("postionData", 60, JSON.stringify(positionEligibilityData));
-    logger.info("position data set to redis");
-   return res
-      .status(200)
-      .send({ message: "list of positions with eligibility", positionEligibilityData });
+
+    const verification = jwt.verify(token, process.env.JWT_KEY);
+    if (!verification) {
+      logger.error("Not verified");
+      return res.status(403).send({ message: "Forbidden" });
+    }
+
+    const positionEligibilityData = await positionEligibilityModel.find({});
+    if (!positionEligibilityData || positionEligibilityData.length <= 0) {
+      return res.status(404).send({ message: "Data not available" });
+    }
+
+    client.setEx("positionData", 60, JSON.stringify(positionEligibilityData));
+    logger.info("Position data set to Redis");
+    
+    return res.status(200).send({
+      message: "List of positions with eligibility",
+      positionEligibilityData,
+    });
   } catch (error) {
-    logger.error("position get error", { error: err });
-    next(err);
-    return res.status(401).send({ message: "data not getting" });
+    logger.error("Error while getting position data", { error: error });
+    return res.status(500).send({ message: "Internal server error" });
   }
 });
+
 /**
  * @swagger
  * /position/{id}:
@@ -119,27 +136,46 @@ positionRoute.get("/position", postionCacheData,async (req, res) => {
  *            description: Internet server problem
  *
  */
-positionRoute.get("/position/:id",particularPositionCache, async (req, res) => {
-  try {
-    let { id } = req.params;
-    const ParticularPositionEligible = await positionEligibilityModel.findOne({_id : id})
-    if(ParticularPositionEligible.length <= 0){
-        return res.status(401).send({ message: " data not available or check id " });
-    }
-    client.setEx(
-      "particularPosition",
-      60,
-      JSON.stringify(ParticularPositionEligible)
-    );
-    logger.info("position data set to redis");
-    return res.status(200).send({ message: " data of this position and eligibility", ParticularPositionEligible });
+positionRoute.get(
+  "/position/:id",
+  
+  async (req, res) => {
+    try {
+      const token = req.headers["authorization"]?.split(" ")[1];
+
+      if (!token) {
+        logger.error("No token provided");
+        return res.status(401).send({ message: "Unauthorized" });
+      }
+      const verification = jwt.verify(token, process.env.JWT_KEY);
+
+      if (!verification) {
+        logger.error("Not verified");
+        return res.status(403).send({ message: "Forbidden" });
+      }
+      let { id } = req.params;
+      const ParticularPositionEligible = await positionEligibilityModel.findOne(
+        { _id: id }
+      );
+      if (ParticularPositionEligible.length <= 0) {
+        return res
+          .status(401)
+          .send({ message: " data not available or check id" });
+      }
     
-   } catch (error) {
-    logger.error("position get error", { error: err });
-      next(err);
-    return res.status(401).send({ message: "data not getting" });
-   }
-});
+      return res.status(200).send({
+        message: " data of this position and eligibility",
+        ParticularPositionEligible,
+      });
+    } catch (error) {
+      logger.error("position get error", { error: error });
+
+      return res
+        .status(500)
+        .send({ message: "Internal Server Error data not getting" });
+    }
+  }
+);
 /**
  * @swagger
  * /positions/{id}:
@@ -172,6 +208,16 @@ positionRoute.get("/position/:id",particularPositionCache, async (req, res) => {
  */
 positionRoute.post("/positions/:id", async (req, res) => {
   try {
+    const token = req.headers["authorization"].split(" ")[1];
+    if (!token) {
+      logger.error("No token provided");
+      return res.status(401).send({ message: "Unauthorized" });
+    }
+    const { role } = jwt.verify(token, process.env.JWT_KEY);
+    if (role !== "Admin") {
+      logger.error("Not authorized");
+      return res.status(403).send({ message: "Not authorized" });
+    }
     const { id } = req.params;
     const {
       title,
@@ -215,22 +261,21 @@ positionRoute.post("/positions/:id", async (req, res) => {
       !twelfthPer ||
       !gender
     ) {
-     return res.status(401).send({ message: "fill all the details" });
+      res.status(401).send({ message: "fill all the details" });
     }
-    if ( typeof openings !== "number" || !Array.isArray(locations) ||
+    if (
+      typeof openings !== "number" ||
+      !Array.isArray(locations) ||
       locations.some((location) => typeof location !== "string")
     ) {
-      logger.info("Invalid input data types");
       return res.status(400).send({ message: "Invalid input data types" });
     }
     const company = await companyData.findOne({ _id: id });
 
     if (!company) {
-      logger.info("Company not found");
       return res.status(404).json({ message: "Company not found" });
     }
-  
-   
+
     const position = new positionEligibilityModel({
       ...req.body,
       companyName: company.companyName,
@@ -242,9 +287,9 @@ positionRoute.post("/positions/:id", async (req, res) => {
     return res
       .status(201)
       .send({ message: "Position save successfully", id: position._id });
-  } catch (err) {
-    console.error(err);
-    res.status(401).json({ message: "data not post" });
+  } catch (error) {
+    logger.error("post position error", { error: error });
+    return res.status(500).send({ message: "Internal Server Error" });
   }
 });
 
@@ -280,16 +325,39 @@ positionRoute.post("/positions/:id", async (req, res) => {
  */
 positionRoute.patch("/updatePosition/:id", async (req, res) => {
   try {
+    const token = req.headers["authorization"].split(" ")[1];
+    if (!token) {
+      logger.error("No token provided");
+      return res.status(401).send({ message: "Unauthorized" });
+    }
+    const { role } = jwt.verify(token, process.env.JWT_KEY);
+    if (role !== "Admin") {
+      logger.error("Not authorized");
+      return res.status(403).send({ message: "Forbidden" });
+    }
+
     let { id } = req.params;
     const updateData = req.body;
-    const Data = await positionEligibilityModel.findByIdAndUpdate(id, updateData, {
-      new: true,
-    });
-    logger.info("position updated successfully");
+    const Data = await positionEligibilityModel.findByIdAndUpdate(
+      id,
+      updateData,
+      {
+        new: true,
+      }
+    );
+
+    // set data to redis //
+    const positionEligibilityData = await positionEligibilityModel.find({});
+
+    if (positionEligibilityData.length <= 0) {
+      return res.status(404).send({ message: "data not available" });
+    }
+    client.setEx("positionData", 60, JSON.stringify(positionEligibilityData));
+    logger.info("position data set to redis");
     return res.status(200).send({ message: "position updated successfully" });
-  } catch (err) {
-    logger.error("position patch with id  error", { error: err });
-   return res.status(401).send({ message: "position updated unsuccessfully" });
+  } catch (error) {
+    console.log(error);
+    return res.status(401).send({ message: "position updated unsuccessfully" });
   }
 });
 
@@ -317,20 +385,35 @@ positionRoute.patch("/updatePosition/:id", async (req, res) => {
  */
 
 positionRoute.delete("/deletePosition/:id", async (req, res) => {
-  let { id } = req.params;
-  const Data = await positionEligibilityModel.findByIdAndDelete({ _id: id });
+  try {
+    const token = req.headers["authorization"].split(" ")[1];
+    if (!token) {
+      logger.error("No token provided");
+      return res.status(401).send({ message: "Unauthorized" });
+    }
+    const { role } = jwt.verify(token, process.env.JWT_KEY);
+    if (role !== "Admin") {
+      logger.error("Not authorized");
+      return res.status(403).send({ message: "Not authorized" });
+    }
+    let { id } = req.params;
+    const Data = await positionEligibilityModel.findByIdAndDelete({ _id: id });
 
-  res.status(200).send({ message: "position deleted successfully", Data });
+    res.status(200).send({ message: "position deleted successfully" });
 
-  // redis set updated data //
-  
-  const positionEligibilityData = await positionEligibilityModel.find({});
+    // redis set updated data //
 
-  if (positionEligibilityData.length <= 0) {
-    return res.status(404).send({ message: "data not available" });
+    const positionEligibilityData = await positionEligibilityModel.find({});
+
+    if (positionEligibilityData.length <= 0) {
+      return res.status(404).send({ message: "data not available" });
+    }
+    client.setEx("positionData", 60, JSON.stringify(positionEligibilityData));
+    logger.info("position data set to redis");
+  } catch (error) {
+    logger.error("delete position error", { error });
+    return res.status(500).send({ message: "Internal server error" });
   }
-  client.setEx("postionData", 60, JSON.stringify(positionEligibilityData));
-  logger.info("position data set to redis");
 });
 
 module.exports = positionRoute;
